finalisation integration solver dans GUI + correctio solver

+++ Core/config.hpp
@@ -7,4 +7,4 @@ const double DIST_MIN_NN = 0.01; //dist minimale non nulle
 const int DIX_P_DECIMALES = 100;
 const double AMELIORATION_MINIMALE = 0.95;
 const int MINIMUM_ITERATIONS = 5;
-const int PI_MAX_SIZE = 0.7;
+const double PI_MAX_SIZE = 0.7;
